backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "djangorestframework-simplejwt>=5.5.0",
    "djoser>=2.3.1",
    "pillow>=11.2.1",
    "psycopg[binary]>=3.2.9",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
]
